Export button and snap to grid button on view page of simulation (#54)
