+++ src/components/controls/children/toggleButtons/AutoSynchSwitchButton.js
@@ -1,0 +1,18 @@
+import {Form} from "react-bootstrap";
+
+const AutoSynchSwitchButton = ({changeAutoSync}) => {
+
+    return <>
+        <Form.Check
+            type="switch"
+            id="auto-sync-switch"
+            label="Autosync"
+            defaultChecked={true}
+            inline={true}
+            onChange={changeAutoSync}
+        />
+    </>
+
+}
+
+export default AutoSynchSwitchButton;
